+++ CHANGELOG.md
@@ -5,7 +5,9 @@ LUYA CHANGELOG
 -------------------------
 
 **[BC BREAK]** removed $app->storage component and replaced with new storage $app->component see api guide. 
+**[BC BREAK]** removed luya\web\Composition::set method.
 
+- `#584` fixed bug where preview link does not handle language correct. Moved language handler into composition component and added new event class.
 - `#582` fixed issue where file downlaod links does not work on multi lingual pages, moved route resolver into urlManager instead of luya urlRule.
 - `#588` delete the block cache (for the specific block) after updating them via cms admin.
 - `#487` added URL duplication check (alias) on page move.